# services/gateway-service/src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Nutri-App Gateway API",
    "version": "1.0.0",
    "description": "Unified API Gateway documentation for all Nutri-App services"
  },
  "servers": [
    {
      "url": "http://localhost:4000/api",
      "description": "Local Gateway Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "LoginRequestDto": {},
      "LoginResponseDto": {},
      "RegisterRequestDto": {},
      "RegisterResponseDto": {},
...
      "LogoutResponseDto": {},
...
      "MfaVerifyRequestDto": {},
      "MfaVerifyResponseDto": {},
...
      "ClientsListResponseDto": {},
...
      "UpsertUserInfoRequestDto": {},
...
    }
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "User login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register new account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Logout current session",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully logged out",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/mfa/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register MFA device",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "MFA device registered"
          }
        }
      }
    },
    "/api/auth/mfa/verify": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Verify MFA code",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaVerifyRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "MFA verified & session created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MfaVerifyResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/clients": {
      "get": {
        "tags": [
          "Clients"
        ],
        "summary": "Get list of enriched client profiles for coach",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of clients",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientsListResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/clients/{id}": {
      "get": {
        "tags": [
          "Clients"
        ],
        "summary": "Get raw client info from IDM",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Client data retrieved"
          }
        }
      }
    },
    "/api/users/{id}/info": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get full user info",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User info retrieved"
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update user info",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertUserInfoRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User info updated"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete user info record",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User info deleted"
          }
        }
      }
    },
...
  },
...
}